Stopgap update for modernization
I've been meaning to overhaul this for years, but it never seems to get done. I figured it's time to cut my losses and at least get all of the obsolete information and defunct links taken down. Hopefully, this will yet be the year that I modernize it, but at least the information's correct again (for now).